% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_summary_table.R
\name{f.cast.tbl}
\alias{f.cast.tbl}
\title{Forecast summary table}
\usage{
f.cast.tbl(f.cast.age, sibling, f.cast.yr, riv.name, start.yr, model)
}
\arguments{
\item{sibling}{A tible with one column}

\item{forecast.age}{A tible with one column}
}
\description{
Forecast summary table
}
\examples{
f.cast.tbl()
}
\keyword{forecast}
