% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_summary_table.R
\name{f.cast.tbl.summary}
\alias{f.cast.tbl.summary}
\title{Summary Performance table}
\usage{
f.cast.tbl.summary(f.cast.age, sibling, f.cast.yr, riv.name, start.yr)
}
\arguments{
\item{sibling}{A tible with one column}

\item{f.cast.yr}{year to forecast (numeric)}

\item{forecast.age}{A tible with one column}

\item{riv.nam}{A character string, river}

\item{start.year}{oldest year to include in the forecast model}
}
\description{
summarizes the performance metrics at 3 and 5 year means
}
\examples{
f.cast.tbl.summary()
}
\author{
Greg Buck
}
\keyword{forecast}
