% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ln_sibling.R
\name{ln.sib.forecast}
\alias{ln.sib.forecast}
\title{log sibling forecast}
\usage{
ln.sib.forecast(sibling, forecast.age, input)
}
\arguments{
\item{sibling}{A tible with one column}

\item{forecast.age}{A tible with one column}
}
\description{
forecast next years log return of a given age class using a linnear regression of the log sibling returns
}
\examples{
ln.sib.forecast()
}
\keyword{log}
\keyword{sibling}
