% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ln_sibling.R
\name{ln.sib.rsquare}
\alias{ln.sib.rsquare}
\title{log sibling R-square}
\usage{
ln.sib.rsquare(sibling, forecast.age)
}
\arguments{
\item{sibling}{A tible with one column}

\item{forecast.age}{A tible with one column}
}
\description{
Returns the R-square value of a linnear regression of log forcasted returns and log sibling returns
}
\examples{
ln.sib.rsquare()
}
\keyword{log}
\keyword{sibling}
