% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnS_lnR.R
\name{lnS.lnR.forecast}
\alias{lnS.lnR.forecast}
\title{log spawner log recruit forecast}
\usage{
lnS.lnR.forecast(spawners, forecast.age, input)
}
\arguments{
\item{forecast.age}{A tible with one column}

\item{sibling}{A tible with one column}
}
\description{
forecast next years log return of a given age class using a linnear regression of the log spawner returns
}
\examples{
lnS.lnR.forecast()
}
\keyword{log}
\keyword{recruit}
\keyword{spawner}
