% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnS_lnR.R
\name{lnS.lnR.pvalue}
\alias{lnS.lnR.pvalue}
\title{lnS.lnR sibling p-value}
\usage{
lnS.lnR.pvalue(spawners, forecast.age)
}
\arguments{
\item{forecast.age}{A tible with one column}

\item{sibling}{A tible with one column}
}
\description{
Returns the slope p-value of a linear regression of the log return of a given age class and the log spawner returns
}
\examples{
lnS.lnR.pvalue()
}
\keyword{log}
\keyword{recruit}
\keyword{spawner}
