% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnS_lnR.R
\name{lnS.lnR.rsquare}
\alias{lnS.lnR.rsquare}
\title{lnS.lnR sibling R-square}
\usage{
lnS.lnR.rsquare(spawners, forecast.age)
}
\arguments{
\item{forecast.age}{A tible with one column}

\item{sibling}{A tible with one column}
}
\description{
Returns the R-square of a linear regression of the log return of a given age class and the log spawner returns
}
\examples{
lnS.lnR.rsquare()
}
\keyword{log}
\keyword{recruit}
\keyword{spawner}
