% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnY_sibling.R
\name{lnY.sib.forecast}
\alias{lnY.sib.forecast}
\title{log Y sibling forecast}
\usage{
lnY.sib.forecast(sibling, forecast.age, input)
}
\arguments{
\item{sibling}{A tible with one column}

\item{forecast.age}{A tible with one column}
}
\description{
forecast next years log return of a given age class using a linnear regression of the sibling returns
}
\examples{
lnY.sib.forecast()
}
\keyword{Y}
\keyword{log}
\keyword{sibling}
