% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnY_sibling.R
\name{lnY.sib.pvalue}
\alias{lnY.sib.pvalue}
\title{log Y sibling R-square value}
\usage{
lnY.sib.pvalue(sibling, forecast.age)
}
\arguments{
\item{sibling}{A tible with one column}

\item{forecast.age}{A tible with one column}
}
\description{
Returns the slope p-value of a linnear regression of log forcasted returns and sibling returns
}
\examples{
lnY.sib.pvalue()
}
\keyword{Y}
\keyword{log}
\keyword{sibling}
