% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_plots.R
\name{plot.data}
\alias{plot.data}
\title{Forecast plots}
\usage{
\method{plot}{data}(f.cast.age, sibling, f.cast.yr, riv.name, start.yr, model)
}
\arguments{
\item{sibling}{A tible with one column}

\item{f.cast.yr}{year to forecast (numeric)}

\item{model}{sibling, ln.sibling, lnY.sibling, lnS.lnR, ricker}

\item{forecast.age}{A tible with one column}

\item{riv.nam}{A character string, river}

\item{start.year}{oldest year to include in the forecast model}
}
\description{
linnear regression plots and residual plot for sibling, ln.sibling, lnY.sibling, lnS.lnR and ricker models
}
\examples{
plot.data()
}
\author{
Greg Buck
}
\keyword{forecast}
