% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ricker.R
\name{ricker.forecast}
\alias{ricker.forecast}
\title{Ricker forecast}
\usage{
ricker.forecast(spawners, forecast.age, input)
}
\arguments{
\item{forecast.age}{A tible with one column}

\item{sibling}{A tible with one column}
}
\description{
forecast next years return of a given age class using a linnear transformation of the Ricker relationship
}
\examples{
ricker.forecast()
}
\keyword{ricker}
