% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ricker.R
\name{ricker.pvalue}
\alias{ricker.pvalue}
\title{Ricker p-value}
\usage{
ricker.pvalue(spawners, forecast.age)
}
\arguments{
\item{forecast.age}{A tible with one column}

\item{sibling}{A tible with one column}
}
\description{
Returns the slope p-value of a linnear transformation of the Ricker relationship
}
\examples{
ricker.pvalue()
}
\keyword{ricker}
