% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ricker.R
\name{ricker.rsquare}
\alias{ricker.rsquare}
\title{Ricker R-square}
\usage{
ricker.rsquare(spawners, forecast.age)
}
\arguments{
\item{forecast.age}{A tible with one column}

\item{sibling}{A tible with one column}
}
\description{
Returns the R-square value of a linnear transformation of the Ricker relationship
}
\examples{
ricker.rsquare()
}
\keyword{ricker}
