% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibling.R
\name{sib.forecast}
\alias{sib.forecast}
\title{sibling forecast}
\usage{
sib.forecast(sibling, forecast.age, input)
}
\arguments{
\item{sibling}{A tible with one column}

\item{forecast.age}{A tible with one column}
}
\description{
forecast next years return of a given age class using a linnear regression of sibling returns
}
\examples{
sib.forecast()
}
\keyword{sibling}
