% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibling.R
\name{sib.pvalue}
\alias{sib.pvalue}
\title{sibling p-value}
\usage{
sib.pvalue(sibling, forecast.age, bb.riv)
}
\arguments{
\item{sibling}{A tible with one column}

\item{forecast.age}{A tible with one column}
}
\description{
Returns the slope p-value of a linear regression of sibling returns
}
\examples{
sib.pvalue()
}
\keyword{sibling}
